[driver] Re-work waiting on IRPs
*sigh* As noted in a previous commit, since some IRPs cannot return
STATUS_PENDING, even with a dedicated thread for a device's IRPs,
there _still_ needs to be [ugly] synchronization.

WvAddIrpToDeviceQueue is now WvlAddIrpToDeviceQueue and takes a
boolean 'Wait' argument so that the caller can either wait for the
IRP to be processed or not.

Soon it'll be the task of a mini-driver's IRP dispatch function to
partially process an IRP and determine which of three cases apply:

  1. Process the IRP completely and asynchronously to other IRPs
  2. Enqueue the IRP for the device's thread and wait for the result
  3. Enqueue the IRP for the device's thread and don't wait

As an example, PnP IRPs would be #2.  As another example, a SCSI
IRP could be either #1 or #3, depending on if PnP IRPs are in the
queue.