[mainbus] More copying from libbus
Copied WvlBusPnpQueryBusInfo content as WvMainBusPnpQueryBusInfo,
with a few changes to allow the main bus' PnP bus info to be set
when the device extension is initialized.

HOWEVER, this doesn't really belong with the main bus PnP handlers,
as it is supposed to be handled by PDOs.  It'll stay here until
it can go in the dummy device source.