[driver,build] Check the Registry for CDDB associations
If there are no CDDB associations, any PDOs that are supposed to be
driven with WinVBlock FDOs will need to have the AddDevice
invoked by the [mini-]driver requesting the PDO, itself.

For example, if the Safe INT 0x13 Hook mini-driver requests a PDO
and one is produced, if Windows does not have a CDDB association
for its device IDs, it'll never be driven with an FDO from the
Safe INT 0x13 Hook mini-driver.

The WvlCddbDone boolean can now be checked by a mini-driver.