[driver] Major work for IRP synchronization
After mainbus catches up, this commit's logic will be unleashed.
Here is the target flow:

An IRP is dispatched to WvDriverDispatchIrp.

[For mini-drivers,] the DriverContext is cleared.  The IRP is passed
to WvDriverAddIrpToDeviceQueueInternal without waiting for the IRP to
be completed.

The device's IRP processing mode is checked.  If it's serial mode,
WvDriverAddIrpToDeviceQueueInternal will be forced to wait and the
IRP will be enqueued for serial processing.  Otherwise, the IRP will
be dispatched to the device's dispatcher.

A device's IRP processing mode will be affected by a call to
WvlWaitForActiveIrps.  This function also waits for all IRPs that came
before it to finish being processed.  When the function returns, there
is no guarantee that the current IRP is the _only_ IRP being processed,
but it _is_ guaranteed that all IRPs arriving _after_ the call returns
are lining up and being processed synchronously until the whole queue
is empty.  Only when the queue is empty will serial mode be disabled.

Thus a PnP IRP can set a mini-driver device extension flag to hold
newly-arriving IRPs, then a call to WvlWaitForActiveIrps will ensure
that the PnP IRP is the only one being processed, upon return.

The means by which an IRP can be "held" is by its processing thread
calling WvlAddIrpToDeviceQueue.  The IRP and other held IRPs _will_
be processed evey so often by the thread (currently 30 seconds),
but they can simply keep re-adding themselves to the queue.
Eventually some new IRP should come along and release them all.

Mini-drivers should use WvlPassIrpUp and WvlPassIrpDown instead of
IoCompleteRequest and IoCallDriver, respectively.

All of the synchronization is done with a whole bunch of interlocked
flags checks, interlocked active-IRP-counts, and a notification
event.  It seems like it should work, though it is a bit difficult
to follow, perhaps. :)