[driver,mainbus,safehook] Safe Hook mini-driver, fixes
Safe hook's probe.c now contains some mini-driver subjects,
such as DriverEntry, AddDevice, Unload routines.

Changed WvSafeHookProbe to WvlCreateSafeHookDevice, which now
is no longer responsible for adding the device to a parent bus.
The caller will do that.

Moved the type definition fot S_WV_SAFEHOOK_PDO into the header
so that probe.c could use it.

Removed an unused declaration from driver.c.

driver.c now invokes the safe hook mini-driver's entry-point.

Fixed WvlCallFunctionInDeviceThread so that if it's called from
inside the device's thread and the caller is waiting, it simply
calls the waited-for function.

Fixed WvDriverAddIrpToDeviceQueueInternal so that if the caller
is waiting, the sentinel value won't be set.  If the caller is
waiting, it is most likely because they are _not_ an IRP being
processed inside the thread and trying to re-enqueue.

Fixed WvMainBusDispatchPnpIrp so that _all_ PnP IRPs are
processed from inside the device's thread and _all_ of them
will wait for outstanding IRPs to complete.  This seems to be
needed.

Hacked WvMainBusPnpRemoveDevice to work with the old "bus nodes"
as well as the new bus relations strategy.

WvMainBusDriveDevice no longer initiates any probe for PDOs.

Hacked WvMainBusInitialBusRelations to work with the old "bus
nodes" as well as the new bus relations strategy.

Cosmetic change in WvMainBusRemoveDevice.

Fixed a bug in WvlRegisterMainBusInitialProbeCallback where a
locked list was being acquired twice.

Some of the hacks for dealing with the old "bus nodes" has to
eventually go away.