[mainbus] Add interim hack to WvDummyAdd_
The semantics for adding a dummy device to a bus will eventually be
that the caller creates a dummy device, then adds it to whatever
bus.  Right now, WvDummyAdd does both and chooses the main bus
unconditionally, but we don't want that.

While things are shuffled around, there will be some hacks just to
make the transition smooth.  It's good if WinVBlock still works
after every commit.

In this one WvDummyAdd_ more resembles what the final
WvlCreateDummyDevice will look like.  It ties all details together
into a single device extension which is freed when the device is
deleted.