[mainbus] Change parameters for WvDummyAdd
It is starting to more closely resemble WvlCreateDevice, because
eventually it will call it.

A caller, such as the safe hook mini-driver, could call this
function to produce a dummy device with the specified IDs _and_
extend that dummy device with additional safe-hook-specific
data.  This is actually the goal, since a safe hook PDO only
has to carry around a note about where the INT 0x13 hook is.

Crazy parameter-checking in this function...  A dummy PDO device
extension will usually be followed by some dummy IDs, which are
always going to be wrapped with an even larger structure of
arbitrary size, and then the caller might want more storage
beyond all that, and it all has to be aligned properly.  Phew.