[driver,mainbus] Fix device tear-down
Remove the NotAvailable member of the WV_S_DEV_EXT structure.
Replace part of its use with a new CvWvlDeviceFlagAvailable flag.

Introduce a new CvWvlDeviceFlagLinked flag.  This flag is set
  - For PDOs: When they are added to a parent bus
  - For FDOs: When they are attached to a device stack

Introduce a new CvWvlDeviceFlagThread flag.  This flag controls
whether or not IRPs and work items can be added to the device's
IRP queue.

The dummy PDO logic now handles a surprise-removal by incrementing
the resource usage for the device and decrementing it during the
subsequent IRP_MN_REMOVE_DEVICE.  This prevents the device's
thread from stopping and deleting the device before the removal
has been completed.  (See below regarding the resource usage.)

Device tear-down now works like this: The device's thread always
has ultimate responsibility for deleting a device.  The thread
will accept IRPs and work items until such a time as:
  - The device is not linked (see explanation above)
  - The device is no longer available due to WvlDeleteDevice
  - The device, as a tracked resource, is only being used by
    the thread

If any of these conditions are false, the thread continues to
service IRPs and work items.  It is up to the IRP dispatcher
to allow PnP IRPs to be processed and to disallow other IRPs.
(A mini-driver might have special PnP handling or other IOCTLs
that it wishes to service, even when a device is on its way
to deletion.)

WvMainBusAddDevice and WvlAttachDeviceToDeviceStack both link
a device.  WvMainBusRemoveDevice and WvlDetachDevice both unlink
a device.