[safehook,mainbus] Switch safe hook over to a mini-driver
Fix a bug in src/winvblock/mainbus/dummyirp.c where such a PDO
should really respond directly to an IRP_MJ_PNP:IRP_MN_START_DEVICE.

Assign a PDO to the safe hook FDO bus in
WvSafeHookPnpQueryDeviceRelations.  That way, it's linked for the
the later PnP probe for its capabilities, and also linked for its
later removal.

Fix a bug in WvSafeHookPnpRemoveDevice where a safe hook FDO would
try to pass WvlDeleteDevice a null pointer.

Add a DummyPdo member to S_WV_SAFE_HOOK_DUMMY_WRAPPER so we can
calculate the offset of the dummy IDs relative to the wrapper's
beginning.

Remove a hack for the old, non-mini-driver safe hook PDOs.

Actually create the safe hook PDO in WvlCreateSafeHookDevice,
instead of it calling the old, non-mini-driver code.