[memdisk] Initial MEMDISK mini-driver
The MEMDISK mini-driver registers itself as a mini-driver,
then registers a main bus initial probe routine.  When that
routine is invoked, it scans low memory for mBFTs.  If it finds
mBFTs, a safe hook PDO is created for each of them.  There is a
hack in place to prevent the later INT 0x13 safe hook walking
from producing duplicates of mBFT safe hooks.

A MEMDISK FDO will drive a safe hook PDO with the MEMDISK
signature and provide, at most, 2 children: The MEMDISK disk
and the previous safe hook in the chain.

A hack is also in place because RAM disks aren't mini-driver
devices, yet.  This also means that removal is temporarily
broken.